requirements.txt
apache-beam[gcp]
apache-airflow
google-cloud-storage
google-cloud-bigquery
apache-airflow-providers-google
apache-airflow-providers-apache-beam
pandas